added error handling in scripts (#160)
* added error handling in scripts

* improvements

* added error handling for new scripts

* optional reboot

* using absolute path for configs

* checking for config files

* absolute path for backend file

* added colours to script logging

* skipping packages update if pyenv is already installed

* new lines before log message for readability

+++ deploy/scripts/install_fonts.sh
@@ -1,12 +1,33 @@
 #!/bin/bash
 
+# Message colours
+ERROR='\033[1;31mERROR:'
+SUCCESS='\033[1;32m'
+WARNING='\033[1;33mWARNING:'
+
 cd ~/
 
-echo "********** Installing Roboto Font **********"
+echo -e "\n${SUCCESS}********** Installing Roboto Font **********\n"
+
 mkdir -p .fonts
-wget -O roboto.zip "https://fonts.google.com/download?family=Roboto"
+
+if [ $( ls .fonts/ | grep -c "Roboto" ) -ge 2 ]
+then
+    echo -e "${WARNING} Roboto fonts already available."
+    exit
+else
+    wget -O roboto.zip "https://fonts.google.com/download?family=Roboto"
+fi
 
 sudo apt install unzip -y
 
-unzip roboto.zip -d .fonts/
-rm roboto.zip
+if [ 1 -eq $( ls | grep -c "roboto.zip" ) ]
+then
+    unzip roboto.zip -d .fonts/
+    rm roboto.zip
+else
+    echo -e "${ERROR} Fonts download failed"
+    exit 1
+fi
+
+echo -e "\n${SUCCESS}Roboto fonts installation complete\n"
